requirements.txt
reportlab
pyinstaller
argparse
Pillow  